requirements.txt
fuse-python
rdflib
requests
structlog
python-json-logger
black
isort
python-magic
pytest
pytest-xdist